+++ CHANGELOG.md
@@ -1,8 +1,9 @@
-## XX.XX.XX
+## 25.1.0
 * Added dynamic resizing functionality for the content zone
+* Added a config option to content (setZoneTimerInterval) to set content zone timer. (Experimental!)
+
 * Improved management of content zone size for better responsiveness
 * Fixed an issue where the build UUID and executable name were missing from crash reports
-* Added a config option to content (setZoneTimerInterval) to set content zone timer. (Experimental!)
 
 ## 24.7.9
 * Improved view tracking capabilities